+++ packages/web/src/common/shortcuts/shortcut.util.tsx
@@ -1,31 +1,8 @@
 import {
-  ArrowDownIcon,
-  ArrowLeftIcon,
-  ArrowRightIcon,
-  ArrowUpIcon,
-  CommandIcon,
-  ControlIcon,
-  type Icon,
-  WindowsLogoIcon,
-} from "@phosphor-icons/react";
-import {
   detectPlatform,
   formatWithLabels,
   resolveModifier,
 } from "@tanstack/react-hotkeys";
-
-// `Meta` is the platform "command" key: ⌘ on macOS, the Windows logo elsewhere.
-const metaIcon: Icon =
-  detectPlatform() === "mac" ? CommandIcon : WindowsLogoIcon;
-
-const keyIconMap: Record<string, Icon> = {
-  Meta: metaIcon,
-  Control: ControlIcon,
-  ArrowUp: ArrowUpIcon,
-  ArrowDown: ArrowDownIcon,
-  ArrowLeft: ArrowLeftIcon,
-  ArrowRight: ArrowRightIcon,
-};
 
 /** Resolves TanStack `Mod` tokens to `Meta` / `Control` for icons and labels. */
 export function expandModInShortcutDisplay(k: string): string {
@@ -37,28 +14,6 @@ export function expandModInShortcutDisplay(k: string): string {
       return part.toLowerCase() === "mod" ? resolvedMod : part;
     })
     .join("+");
-}
-
-export function ShortCutLabel({ k, size = 14 }: { k: string; size?: number }) {
-  const display = expandModInShortcutDisplay(k);
-
-  return display.split("+").map((_key) => {
-    const key = _key.trim();
-    const testId = `${key.toLowerCase()}-icon`;
-    const IconComponent = keyIconMap[key];
-
-    if (IconComponent) {
-      return <IconComponent key={key} size={size} data-testid={testId} />;
-    }
-
-    // Text keys inherit the surrounding font size (e.g. the keycap chip's 11px)
-    // so letters and modifier icons read consistently together.
-    return (
-      <span key={key} data-testid={testId}>
-        {key}
-      </span>
-    );
-  });
 }
 
 /**
