+++ packages/web/src/components/SelectView/SelectView.test.tsx
@@ -185,7 +185,6 @@ describe("SelectView", () => {
 
       const dayOption = screen.getByRole("option", { name: /day/i });
       expect(dayOption).toHaveAttribute("aria-selected", "true");
-      expect(dayOption).toHaveClass("text-accent-primary");
 
       const weekOption = screen.getByRole("option", { name: /week/i });
       expect(weekOption).toHaveAttribute("aria-selected", "false");
@@ -287,7 +286,8 @@ describe("SelectView", () => {
 
       await waitFor(() => {
         const weekOption = screen.getByRole("option", { name: /week/i });
-        expect(weekOption.className).toContain("bg-white/10");
+        expect(weekOption).toHaveAttribute("tabindex", "0");
+        expect(dayOption).toHaveAttribute("tabindex", "-1");
       });
     });
 
@@ -303,7 +303,8 @@ describe("SelectView", () => {
 
       await waitFor(() => {
         const weekOption = screen.getByRole("option", { name: /week/i });
-        expect(weekOption.className).toContain("bg-white/10");
+        expect(weekOption).toHaveAttribute("tabindex", "0");
+        expect(dayOption).toHaveAttribute("tabindex", "-1");
       });
     });
 
@@ -348,13 +349,13 @@ describe("SelectView", () => {
       dayOption.focus();
 
       expect(dayOption).toHaveAttribute("aria-selected", "true");
-      expect(dayOption).toHaveClass("text-accent-primary");
 
       await user.keyboard("{ArrowDown}");
 
       await waitFor(() => {
         const weekOption = screen.getByRole("option", { name: /week/i });
-        expect(weekOption.className).toContain("bg-white/10");
+        expect(weekOption).toHaveAttribute("tabindex", "0");
+        expect(dayOption).toHaveAttribute("tabindex", "-1");
       });
     });
 
@@ -370,7 +371,8 @@ describe("SelectView", () => {
 
       await waitFor(() => {
         const dayOption = screen.getByRole("option", { name: /day/i });
-        expect(dayOption.className).toContain("bg-white/10");
+        expect(dayOption).toHaveAttribute("tabindex", "0");
+        expect(weekOption).toHaveAttribute("tabindex", "-1");
       });
     });
 
@@ -386,7 +388,8 @@ describe("SelectView", () => {
 
       await waitFor(() => {
         const weekOption = screen.getByRole("option", { name: /week/i });
-        expect(weekOption.className).toContain("bg-white/10");
+        expect(weekOption).toHaveAttribute("tabindex", "0");
+        expect(dayOption).toHaveAttribute("tabindex", "-1");
       });
     });
   });
