+++ packages/web/src/views/Day/components/Toasts/UndoToast/UndoDeleteToast.test.tsx
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, mock } from "bun:test";
 import "@testing-library/jest-dom";
 import { fireEvent, render, screen } from "@testing-library/react";
-import { getModifierKeyTestId } from "@web/common/utils/shortcut/shortcut.util";
+import { getModifierKeyTestId } from "@web/common/shortcuts/shortcut.util";
 
 type ToastMock = ReturnType<typeof mock> & {
   dismiss: ReturnType<typeof mock>;
